requirements.txt
streamlit
google-generativeai
python-dotenv
langchain_core
langchain_community
